refactor(indexer/rpc): extract pool-state fetchers + caches + test mocks to rpc/pool-state.ts (PR-S8)
Move lines 46-911 of rpc.ts (~865 lines) into the new
indexer-envio/src/rpc/pool-state.ts:
- Pool/reserves/decimals/fees test mocks + helpers
- Block-scoped caches (numReporters, reportExpiry, reserves) +
  evictCacheForChain
- RebalancingState type + cache
- All eleven fetchers: fetchRebalancingState, fetchReserves,
  fetchInvertRateFeed, fetchRebalanceThreshold, fetchReferenceRateFeedID,
  fetchNumReporters, fetchReportExpiry, fetchTokenDecimalsScaling,
  fetchTradingLimits, fetchRebalanceIncentiveAtBlock, fetchFees

The -1/-2 sentinel contract on every getter (especially fetchFees'
Promise.allSettled per-getter pattern that pool.ts:438 retry depends on)
is preserved verbatim. The Oracle DB query helpers (lines 912+) and
Breaker RPC self-heal (line 970+) stay in rpc.ts for PR-S9.

rpc.ts re-exports the previously-public surface so feeToken.ts,
EventHandlers.ts, breakers.ts, pool.ts, plus the test files keep
working without import-path changes.

Pure extraction. 380 tests pass.
